requirements.txt
# Ribit 2.0 - Python Dependencies

# Matrix Client
matrix-nio[e2e]>=0.20.0

# DeltaChat Integration
deltabot>=0.8.0

# HTTP and Web
aiohttp>=3.8.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Wikipedia
wikipedia-api>=0.5.8

# Image Processing
Pillow>=9.0.0

# File Type Detection
python-magic>=0.4.27

# Async File Operations
aiofiles>=23.0.0

# Supabase for bridge state management
supabase>=2.0.0

# Async Support (usually built-in, but listed for clarity)
asyncio; python_version < '3.7'

# Optional ROS support
# rclpy; python_version >= '3.7'
# rospy; python_version >= '3.7'

# Image Processing (for offline image analysis)
numpy>=1.24.0
opencv-python-headless>=4.8.0

# Process Management (for bot control)
psutil>=5.9.0
nio
